Re-making a commit for changing data format --kazu

+++ UserDev/CMTool/ClusterMerger.cc
@@ -53,11 +53,11 @@ namespace larlight {
     out_cluster_v.set_run(ev_cluster->run());
     out_cluster_v.set_subrun(ev_cluster->subrun());
     //tmp_index to know what plane to use
-    unsigned short tmp_index = 0;
+    unsigned int tmp_index = 0;
 
     for(auto const& indexes : merged_indexes) {
 
-      std::vector<unsigned short> merged_association;
+      std::vector<unsigned int> merged_association;
 
       for(auto const& cluster_index : indexes) {
